Changes logic for checking when a spike trap should be run into/ removed
Previously, we were looping through all the spike trap states, and making the assumption that the state would match up with the specific spike trap. This is a bad assumption if we have more then one spike trap in the world.

Now, we find the specific spike trap state that matches with the current spike trap. This change should prevent crashes and bad behaviour when too many spike traps are deployed into the world